Added sfx to the bow and arrows.
I think it's pretty messy, could definetely improve on making it easier to add more Item SFX.
Added a script ItemSFX that at the moment checks the type of weapon, and plays and stops the sound of the bow.
For the arrows, added an AudioSource that plays a whoosh on awake, and then a hit sound when it collides.